day14 part 1

+++ src/test/java/fr/game/advent/day14/GameOneTest.java
@@ -116,6 +116,6 @@ public class GameOneTest {
 
 	@Test
 	public void testGame() {
-		Assert.assertEquals(new Integer(0), gameOne.play());
+		Assert.assertEquals(new Integer(502491), gameOne.play());
 	}
 }
